Added support for user based placeholder values in pam_oath usersfile string.
These changes introduce the the "${USER}" and "${HOME}" placeholder
values for the "usersfile" string in the pam_oath configuration
file. The placeholder values allow the user credentials file to be
stored in a file path that is relative to the user, and mimicks
similar behavior found in google-authenticator-libpam.

The motivation for these changes is to allow for non-privileged
processes to use pam_oath (e.g., for 2FA with
xscreensaver). Non-priviledged and non-suid programs are unable to use
pam_oath. These changes are a proposed alternative to a suid helper
binary as well.

The user credential file ownership is automatically adjusted if
"${USER}" and/or "${HOME}" is used. The default is the owner of the
calling process (as returned by getuid()), but if either placeholder
is used the UID of the user is obtained from
pam_modutil_getpwnam(). The ownership is changed just prior the
renaming the new modified user crediential file, while the lock is in
place.